Просто проделана большая работа

+++ app/src/main/java/com/example/weatherApp/RestServices/DataParseClassForCurrentDay.kt
@@ -36,7 +36,7 @@ data class DataParseClassForCurrentDay(
 
     data class Wind(
         @SerializedName("speed") val speed: Double,
-        @SerializedName("deg") val deg: Int
+        @SerializedName("deg") val deg: Double
     )
 
     data class Main(
